+++ locksmith/src/operations/keysOperations.ts
@@ -1,8 +1,21 @@
-import { Members } from '../graphql/datasource'
+import { Keys } from '../graphql/datasource'
 import { Web3Service } from '@unlock-protocol/unlock-js'
 import networks from '@unlock-protocol/networks'
 import * as metadataOperations from './metadataOperations'
 import Fuse from 'fuse.js'
+
+interface Lock {
+  keys: {
+    owner: {
+      address: string
+    }
+    keyId: string
+    expiration: string
+  }[]
+  address: string
+  name: string
+  owner: string
+}
 
 const KEY_FILTER_MAPPING: { [key: string]: string } = {
   owner: 'keyholderAddress',
@@ -10,17 +23,17 @@ const KEY_FILTER_MAPPING: { [key: string]: string } = {
   email: 'email',
 }
 /**
- * Filters members base on query
- * @param {Array} members - list of members
+ * Filters keys base on query
+ * @param {Array} keys - list of keys
  * @param {String} query - query to use as filter on items
- * @return {Array} - list of filtred members by query
+ * @return {Array} - list of filtred keys by query
  */
-async function filterMembers(members: any[], filters: any) {
+async function filterKeys(keys: any[], filters: any) {
   const { query, filterKey } = filters
   const searchByCheckInTime = filterKey === 'checkedInAt'
-  if (!query?.length && !searchByCheckInTime) return members
+  if (!query?.length && !searchByCheckInTime) return keys
 
-  const fuse = new Fuse(members, {
+  const fuse = new Fuse(keys, {
     threshold: 0,
     ignoreLocation: true,
     keys: [KEY_FILTER_MAPPING[filterKey] ?? filterKey],
@@ -35,7 +48,38 @@ async function filterMembers(members: any[], filters: any) {
   })
 }
 
-export async function getMembersWithMedata({
+/** merge keys items with the corresponding metadata value */
+export const buildKeysWithMetadata = (lock: Lock, metadataItems: any[]) => {
+  return lock?.keys
+    ?.map((key: any) => {
+      // get key metadata for the owner
+      const { userMetadata, extraMetadata } =
+        metadataItems?.find(
+          (metadata) =>
+            metadata?.userAddress?.toLowerCase() ===
+            key?.owner?.address?.toLowerCase()
+        )?.data ?? {}
+
+      const metadata = {
+        ...userMetadata?.private,
+        ...userMetadata?.protected,
+        ...extraMetadata,
+      }
+
+      const merged = {
+        token: key?.keyId,
+        lockName: lock?.name,
+        expiration: key?.expiration,
+        keyholderAddress: key?.owner?.address,
+        lockAddress: lock?.address,
+        ...metadata,
+      }
+      return merged
+    })
+    .filter(Boolean)
+}
+
+export async function getKeysWithMetadata({
   network,
   lockAddress,
   filters,
@@ -54,7 +98,7 @@ export async function getMembersWithMedata({
   )
 
   let metadataItems = []
-  const client = new Members(network)
+  const client = new Keys(network)
 
   const [lock] = await client.get({
     addresses: [lockAddress],
@@ -70,10 +114,7 @@ export async function getMembersWithMedata({
     })
   }
 
-  const members = metadataOperations.buildMembersWithMetadata(
-    lock,
-    metadataItems
-  )
+  const keys = buildKeysWithMetadata(lock, metadataItems)
 
-  return filterMembers(members, filters)
+  return filterKeys(keys, filters)
 }
