+++ locksmith/src/operations/metadataOperations.ts
@@ -8,10 +8,7 @@ import networks from '@unlock-protocol/networks'
 import { Verifier } from '../models/verifier'
 import Normalizer from '../utils/normalizer'
 import * as lockOperations from './lockOperations'
-const models = require('../models')
 const Asset = require('../utils/assets')
-
-const { Lock } = models
 
 const baseURIFragement = 'https://assets.unlock-protocol.com'
 
@@ -20,19 +17,6 @@ interface IsKeyOrLockOwnerOptions {
   lockAddress: string
   keyId: string
   network: number
-}
-
-interface Lock {
-  keys: {
-    owner: {
-      address: string
-    }
-    keyId: string
-    expiration: string
-  }[]
-  address: string
-  name: string
-  owner: string
 }
 
 export const updateKeyMetadata = async (data: any) => {
@@ -255,35 +239,4 @@ export const getKeysMetadata = async ({
 
   const mergedData = await Promise.all(mergedDataList)
   return mergedData.filter(Boolean)
-}
-
-/** merge keys items with the corresponding metadata value */
-export const buildMembersWithMetadata = (lock: Lock, metadataItems: any[]) => {
-  return lock?.keys
-    ?.map((key: any) => {
-      // get key metadata for the owner
-      const { userMetadata, extraMetadata } =
-        metadataItems?.find(
-          (metadata) =>
-            metadata?.userAddress?.toLowerCase() ===
-            key?.owner?.address?.toLowerCase()
-        )?.data ?? {}
-
-      const metadata = {
-        ...userMetadata?.private,
-        ...userMetadata?.protected,
-        ...extraMetadata,
-      }
-
-      const merged = {
-        token: key?.keyId,
-        lockName: lock?.name,
-        expiration: key?.expiration,
-        keyholderAddress: key?.owner?.address,
-        lockAddress: lock?.address,
-        ...metadata,
-      }
-      return merged
-    })
-    .filter(Boolean)
 }
