+++ locksmith/src/routes/v2/key.ts
@@ -1,6 +1,6 @@
 import express from 'express'
 import KeyController from '../../controllers/v2/keyController'
-import { authenticatedMiddleware } from '../../utils/middlewares/auth'
+//import { authenticatedMiddleware } from '../../utils/middlewares/auth'
 
 const router = express.Router({ mergeParams: true })
 
@@ -8,7 +8,7 @@ const memberController = new KeyController()
 
 router.get(
   '/:network/locks/:lockAddress/keys',
-  authenticatedMiddleware,
+  //authenticatedMiddleware,
   (req, res) => {
     memberController.keys(req, res)
   }
